backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Document processing
python-docx==1.1.0
pypdf==3.17.4

# Mistral AI
mistralai>=1.0.0

# Data validation
pydantic>=2.8.2,<2.9.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0
